backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==6.4
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
faiss-cpu==1.9.0
fastapi==0.115.5
fastapi-another-jwt-auth==0.1.11
fastapi-cli==0.0.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jinja2==3.1.4
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.4
langchain-core==0.3.18
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
mako==1.3.6
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.54.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
propcache==0.2.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.1
pygments==2.18.0
pyjwt==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruff==0.7.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.0
typer==0.13.0
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
werkzeug==3.1.3
yarl==1.17.1
